Tried to add the Board creation to the FileReader. There is a problem with the copy operations.

+++ Battleship2/src/Scores.cpp
@@ -11,7 +11,7 @@ int Scores::activeThreads = 0;
 
 
 
-void Scores::initScores(int numberOfPlayers)
+void Scores::initScores(size_t numberOfPlayers)
 {
 	for (int i = 0; i < numberOfPlayers; i++)
 	{
